Drugi dzień TESTÓW

+++ src/main/java/pl/sdacademy/animals/bear/BlackBear.java
@@ -1,5 +1,9 @@
 package pl.sdacademy.animals.bear;
 
+import org.joda.time.DateTime;
+import org.joda.time.Duration;
+import org.joda.time.LocalDate;
+import org.joda.time.LocalDateTime;
 import pl.sdacademy.clock.Clock;
 
 public class BlackBear extends Bear {
@@ -7,7 +11,11 @@ public class BlackBear extends Bear {
         super(weight, clock);
     }
 
-    public BlackBear(int weight) {
+    public BlackBear(double weight) {
         super(weight);
     }
+
+//    public boolean isHibernating() {
+//        boolean result = DateTime.
+//    }
 }
